listeners and publishers config

+++ src/main/java/com/example/gestionAcueducto/users/service/Impl/PasswordForgotServiceImpl.java
@@ -1,11 +1,9 @@
 package com.example.gestionAcueducto.users.service.Impl;
 
-import com.example.gestionAcueducto.events.ResetPasswordRequestEvent;
 import com.example.gestionAcueducto.users.dto.PasswordForgotRequest;
 import com.example.gestionAcueducto.users.dto.SimpleMessageDTO;
 import com.example.gestionAcueducto.users.entity.PasswordResetToken;
 import com.example.gestionAcueducto.users.entity.User;
-import com.example.gestionAcueducto.users.service.UserMessageService;
 import com.example.gestionAcueducto.users.service.PasswordForgotService;
 import com.example.gestionAcueducto.users.service.PasswordResetTokenService;
 import com.example.gestionAcueducto.users.service.UserService;
@@ -21,7 +19,6 @@ public class PasswordForgotServiceImpl implements PasswordForgotService {
 
     private final UserService userService;
     private final PasswordResetTokenService passwordResetTokenService;
-    private final UserMessageService userMessageService;
 
 
 
@@ -32,27 +29,13 @@ public class PasswordForgotServiceImpl implements PasswordForgotService {
         PasswordResetToken passwordResetToken;
         passwordResetToken = passwordResetTokenService.findByUser(user);
 
-        if(passwordResetToken != null){
-
-            if(!passwordResetTokenService.isTokenExpired(passwordResetToken)){
-
-                ResetPasswordRequestEvent resetPasswordRequestEvent = new ResetPasswordRequestEvent( passwordResetToken.getSagaId(),
-                        user.getEmail(), "reset-password-email", passwordResetToken.getToken());
-
-                userMessageService.publishResetPasswordEvent(resetPasswordRequestEvent);
-
-                return new SimpleMessageDTO("Se ha reenviado el correo exitosamente!");
-            }
+        if(passwordResetToken != null && passwordResetTokenService.isTokenExpired(passwordResetToken)){
             passwordResetTokenService.deleteToken(passwordResetToken);
         }
 
-        passwordResetToken = passwordResetTokenService.createPasswordResetToken(user, 5);
 
+        passwordResetTokenService.createPasswordResetToken(user, 5);
 
-        ResetPasswordRequestEvent resetPasswordRequestEvent = new ResetPasswordRequestEvent( passwordResetToken.getSagaId(),
-                user.getEmail(), "reset-password-email", passwordResetToken.getToken());
-
-        userMessageService.publishResetPasswordEvent(resetPasswordRequestEvent);
 
         return new SimpleMessageDTO("Se ha generado un nuevo correo de recuperación exitosamente!");
